Remove db encryption with pgcrypto
