+++ horreum-backend/src/main/resources/application.properties
@@ -10,6 +10,11 @@ quarkus.datasource.jdbc.min-size=3
 quarkus.datasource.jdbc.max-size=25
 quarkus.datasource.jdbc.initial-size=3
 
+## see https://jdbc.postgresql.org/documentation/ssl/ to configure SSL on the driver and the server
+%prod.quarkus.datasource.jdbc.additional-jdbc-properties.ssl=true
+%prod.quarkus.datasource.jdbc.additional-jdbc-properties.sslmode=verify-full
+##%prod.quarkus.datasource.jdbc.additional-jdbc-properties.sslrootcert=server.crt
+
 # amqp
 %prod.amqp-host=localhost
 %prod.amqp-port=5672
