+++ horreum-integration-tests/src/main/resources/application.properties
@@ -1,2 +1,5 @@
 quarkus.quinoa=false
 quarkus.container-image.build=false
+
+## disable certificate validation (but still require a SSL connection)
+quarkus.datasource.jdbc.additional-jdbc-properties.sslmode=require
